refactor(added index.js file): provided the support of default export and multiple exports
