+++ __app/component/TextToSpeech/index.js
@@ -1,15 +1,15 @@
-import TextToSpeechInit from './TextToSpeechInit';
+import TextToSpeech from './TextToSpeech';
 import TextToSpeechStart from './TextToSpeechStart';
 import TextToSpeechStop from './TextToSpeechStop';
 
 export {
-  TextToSpeechInit,
+  TextToSpeech,
   TextToSpeechStart,
   TextToSpeechStop,
 };
 
 export default {
-  Init: TextToSpeechInit,
+  Init: TextToSpeech,
   Start: TextToSpeechStart,
   Stop: TextToSpeechStop,
 };
