+++ __app/component/VoiceRecognition/VoiceRecognition.js
